requirements.txt
flask
flask-sqlalchemy
flask-wtf
flask-migrate
python-dotenv
pillow
gunicorn
psycopg2-binary
